// src/types/resources/CarePlan.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCanonical } from '../simpleTypes/Canonical';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TPeriod } from '../partials/Period';
import { TDateTime } from '../simpleTypes/DateTime';
import { TCarePlanActivity } from '../partials/CarePlanActivity';
import { TAnnotation } from '../partials/Annotation';

export type TCarePlan = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  instantiatesCanonical?: TCanonical[];
  instantiatesUri?: TUri[];
  basedOn?: TReference[];
  replaces?: TReference[];
  partOf?: TReference[];
  status: string;
  intent: string;
  category?: TCodeableConcept[];
  title?: string;
  description?: string;
  subject: TReference;
  encounter?: TReference;
  period?: TPeriod;
  created?: TDateTime;
  author?: TReference;
  contributor?: TReference[];
  careTeam?: TReference[];
  addresses?: TReference[];
  supportingInfo?: TReference[];
  goal?: TReference[];
  activity?: TCarePlanActivity[];
  note?: TAnnotation[];
};